+++ src/main/kotlin/org/developerden/codosseum/validation/TemplateValidation.kt
@@ -1,7 +1,7 @@
 package org.developerden.codosseum.validation
 
 import io.ktor.client.call.*
-import org.developerden.codosseum.indexing.challenge.Challenge
+import org.developerden.codosseum.indexing.Challenge
 import org.developerden.codosseum.templatespiler.api.apis.DefaultApi
 import org.developerden.codosseum.templatespiler.api.models.TemplateParseRequest
